extracted Logger::Application from logger.rb
